requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-multipart
python-jose
aiofiles
requests
pytest
jinja2
websockets
email-validator
Faker
face_recognition
opencv-python
python-dotenv
pymysql
python-dotenv
